Reorganize file structure of itest cases
Changes:
* Group tests by shared semantics, rather than having all on top level.
* Remove 2 tests in `builtin.rs` (Vector2, VariantArray), already covered through unit-tests and other itests.
