+++ itest/rust/src/builtin_tests/color_test.rs
