+++ itest/rust/src/builtin_tests/containers/array_test.rs
