+++ itest/rust/src/builtin_tests/geometry/rect2_test.rs
