+++ itest/rust/src/builtin_tests/geometry/rect2i_test.rs
