+++ itest/rust/src/builtin_tests/mod.rs
@@ -1,0 +1,33 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+mod geometry {
+    mod basis_test;
+    mod plane_test;
+    mod projection_test;
+    mod quaternion_test;
+    mod rect2_test;
+    mod rect2i_test;
+    mod transform2d_test;
+    mod transform3d_test;
+}
+
+mod containers {
+    mod array_test;
+    mod callable_test;
+    mod dictionary_test;
+    mod rid_test;
+    mod signal_test;
+    mod variant_test;
+}
+
+mod string {
+    mod godot_string_test;
+    mod node_path_test;
+    mod string_name_test;
+}
+
+mod color_test;
