+++ itest/rust/src/builtin_tests/string/node_path_test.rs
