+++ itest/rust/src/builtin_tests/string/string_name_test.rs
