+++ itest/rust/src/engine_tests/native_structures_test.rs
