+++ itest/rust/src/object_tests/enum_test.rs
