+++ itest/rust/src/object_tests/property_test.rs
