+++ itest/rust/src/object_tests/singleton_test.rs
