+++ itest/rust/src/object_tests/virtual_methods_test.rs
