+++ itest/rust/src/register_tests/constant_test.rs
