+++ itest/rust/src/register_tests/derive_variant/enums_skip.rs
