+++ itest/rust/src/register_tests/derive_variant/structs_skip.rs
