+++ itest/rust/src/register_tests/mod.rs
@@ -1,0 +1,12 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+mod constant_test;
+mod derive_variant;
+mod func_test;
+mod gdscript_ffi_test;
+mod option_ffi_test;
+mod var_test;
